requirements.txt
streamlit>=1.34.0
google-genai>=0.5.0
youtube-transcript-api
yt-dlp
opencv-python-headless
pillow
python-dotenv
